+++ src/components/Common/TextLink/index.tsx
@@ -1,17 +1,27 @@
-import React, { FC } from 'react';
+import React, { FC, MouseEvent } from 'react';
 import { colors, m, StyleProps, Styles } from 'styles';
 
 interface Props extends StyleProps {
   href?: string;
   target?: string;
   className?: string;
+  onClick?: () => void;
 }
 
-export const TextLink: FC<Props> = ({ style, href = '#', target, className, children }) => (
-  <a className={className} style={m([styles.container, style])} href={href} target={target}>
-    {children}
-  </a>
-);
+export const TextLink: FC<Props> = ({ style, href = '#', target, className, children, onClick }) => {
+  const handleClick = (e: MouseEvent<HTMLAnchorElement>) => {
+    if (onClick) {
+      e.preventDefault();
+      onClick();
+    }
+  };
+
+  return (
+    <a className={className} style={m([styles.container, style])} href={href} target={target} onClick={handleClick}>
+      {children}
+    </a>
+  );
+};
 
 const styles: Styles = {
   container: {