+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.13] - 2021-02-24
+### Changed
+- Signup confirmation message changed
+
+### Fixed
+- Email confirmation autofill fixed
+
 ## [1.0.12] - 2021-02-23
 ### Added
 - DO-36: Forgot Password functionality implemented